Moved dereference node utils to ast.ts
